+++ routes/signin.tsx
@@ -2,6 +2,7 @@ import { Header, Input, Text } from 'lunchbox';
 import { Handlers, PageProps } from '$fresh/server.ts';
 import { redirect } from 'redirect';
 import { WithSession } from 'fresh_session';
+import { firstSignInSetup, getAppConfiguration } from 'setup';
 
 type Data = {
   session: Record<string, string>;
@@ -22,6 +23,7 @@ export const handler: Handlers<Data, WithSession> = {
 
     if (enteredPassword === Deno.env.get('PASSWORD')) {
       await ctx.state.session.set('isSignedIn', true);
+      await firstSignInSetup();
 
       return redirect('/');
     } else {